% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_update.R
\name{beta_update}
\alias{beta_update}
\title{Update the vector.}
\usage{
beta_update(x, y, beta, lambda, lambda0)
}
\arguments{
\item{x}{The predictor variable in the elastic net(E-N) problem.}

\item{y}{The response variable in the E-N problem.}

\item{beta}{The initial value of the vector to update.}

\item{lambda}{The coefficient of the penalty term in the E-N problem.}

\item{lambda0}{The coefficient of the penalty term L2 in the E-N problem.}
}
\description{
Use coordinate descent method to update the vector. The formula is derived from the derivation of single variable problem and the discussion of simple classification. The output is the vector that has been updated.
}
