% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMC}
\alias{EMC}
\title{EM of certain case}
\usage{
EMC(lambda1, lambda2, u, v)
}
\arguments{
\item{lambda1}{This method requires two initialization parameters, it's the first one.}

\item{lambda2}{This method requires two initialization parameters, it's the second one.}

\item{u}{A vector consisting of the left bounds of several known intervals.}

\item{v}{A vector consisting of the right bounds of several known intervals.}
}
\value{
The estimate of a from the EM method.
}
\description{
EM method of certain case: we only know that Xi's have distribution Exp(a)(i.i.d) and Xi falls in (ui,vi) for i=1,...,n. This function can use to get the unknown parameter a.
}
\examples{
\dontrun{
lambda1 <- 2
lambda2 <- 1
u <- c(11,8,27,13,16,0,23,10,24,2)
v <- c(12,9,28,14,17,1,24,11,25,3)
emc <- EMC(lambda1,lambda2,u,v)
emc
}
}
