% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPCA.R
\name{SPCA}
\alias{SPCA}
\title{SPCA}
\usage{
SPCA(cormat, k, eps, lambda, lambda1)
}
\arguments{
\item{cormat}{The correlation coefficient matrix of the original data (if only the original data, then R function cor can be used).}

\item{k}{The number of principal components to select.}

\item{eps}{A small value used to control convergence.}

\item{lambda}{L2 penalty term coefficient.}

\item{lambda1}{L1 penalty term coefficient vector.}
}
\description{
PCA is widely used in data processing and dimensionality reduction. However, because the principal component obtained by the method is a linear combination of all the predictive variables, it is difficult to use such principal components to explain the experimental results in many cases. We wanted to filter out the more important (explained with a larger proportion of variance) variables. We need to make the selected principal component as sparse as possible while ensuring that the proportion of variance that can be explained by it is large enough. The SPCA can implement this idea. The idea of SPCA is to transform PCA into a regression-type optimization problem and obtain sparse coefficient matrix by applying lasso constraint to regression coefficient. The output are three items. Loadings: the Loading obtained by SPCA method. Pev: The proportion of variance explained by the principal component. Cumsum_pev: The cumulative proportion of variance explained by the principal component.
}
\examples{
\dontrun{
library(elasticnet)
data(pitprops)
SPCA(pitprops,6,1e-6,0,c(0.06,0.16,0.1,0.5,0.5,0.5))
}
}
